src/requirements.txt
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.40.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
idna==2.8
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.4
Pygments==2.5.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-core==3.2.0
sqlparse==0.3.0
urllib3==1.25.7
wincertstore==0.2
